Use babel, remove InfoListItem PropTypes, new types file
