+++ components/.babelrc
@@ -1,0 +1,11 @@
+{
+  "presets": [
+    "@babel/react",
+    "@babel/env",
+    "@babel/typescript"
+  ],
+  "plugins": [
+    "@babel/proposal-class-properties",
+    "babel-plugin-typescript-to-proptypes"
+  ]
+}
