+++ components/package.json
@@ -4,7 +4,11 @@
   "description": "React components for PX Blue applications",
   "scripts": {
     "test": "jest src",
-    "build": "tsc --p tsconfig.lib.json",
+    "type-check": "tsc --noEmit",
+    "build:original": "tsc --p tsconfig.lib.json",
+    "build": "npm run build:types && npm run build:js",
+    "build:types": "tsc --p tsconfig.lib.json --emitDeclarationOnly",
+    "build:js": "babel src --out-dir ../dist --extensions \".ts,.tsx\"",
     "watch:storybook": "tsc --p tsconfig.lib.json --watch --outdir ../demos/storybook/node_modules/@pxblue/react-components",
     "lint": "eslint \"src/**/**.{tsx,ts}\"",
     "lint:fix": "eslint \"src/**/**.{tsx,ts}\" --fix",
@@ -16,6 +20,12 @@
     "color": "^3.1.2"
   },
   "devDependencies": {
+    "@babel/cli": "^7.8.3",
+    "@babel/core": "^7.8.3",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-react": "^7.9.0",
+    "@babel/preset-typescript": "^7.8.3",
     "@material-ui/core": "^4.0.0",
     "@material-ui/icons": "^3.0.2",
     "@pxblue/eslint-config": "1.1.4",
@@ -28,6 +38,7 @@
     "@types/react-dom": "^16.9.4",
     "@typescript-eslint/eslint-plugin": "^2.16.0",
     "@typescript-eslint/parser": "^2.16.0",
+    "babel-plugin-typescript-to-proptypes": "^1.3.2",
     "clsx": "^1.1.0",
     "enzyme": "^3.5.0",
     "enzyme-adapter-react-16": "^1.8.0",
@@ -43,9 +54,9 @@
   },
   "peerDependencies": {
     "@material-ui/core": "^4.7.2",
+    "@pxblue/colors": "^2.0.0",
     "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "@pxblue/colors": "^2.0.0"
+    "react-dom": "^16.8.6"
   },
   "repository": {
     "type": "git",
